# docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  models.Aluno:
    properties:
      cpf:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      nome:
        type: string
      rg:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
paths:
  /alunos:
    get:
      consumes:
      - application/json
      description: Exibe Alunos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Aluno'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Exibe Alunos
      tags:
      - Alunos
    post:
      consumes:
      - application/json
      description: Cria Alunos
      parameters:
      - description: Body para criar um aluno
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Aluno'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aluno'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Cria Alunos
      tags:
      - Alunos
  /alunos/{id}:
    delete:
      consumes:
      - application/json
      description: Deleta Alunos
      parameters:
      - description: Aluno ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aluno'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Deleta Alunos
      tags:
      - Alunos
    get:
      consumes:
      - application/json
      description: Exibe Alunos por ID
      parameters:
      - description: Aluno ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aluno'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Exibe Alunos por ID
      tags:
      - Alunos
    patch:
      consumes:
      - application/json
      description: Edita Alunos
      parameters:
      - description: Aluno ID
        in: path
        name: id
        required: true
        type: integer
      - description: O body para editar um aluno
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Aluno'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aluno'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Edita Alunos
      tags:
      - Alunos
  /alunos/cpf/{cpf}:
    get:
      consumes:
      - application/json
      description: Exibe Alunos por CPF
      parameters:
      - description: CPF do aluno
        in: path
        name: cpf
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aluno'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Exibe Alunos por CPF
      tags:
      - Alunos
swagger: "2.0"